Move authorization revocation to background jobs
